Cherry pick jsonapi serialzier modules :walking:
We don't need to import jsonapi error module since we have our own implementation of it
