+++ test/__mocks__/deserializer.js
@@ -1,0 +1,11 @@
+var Deserializer = jest.fn();
+
+Deserializer.prototype.deserialize = jest.fn().mockReturnValue(
+  Promise.resolve({
+    id: '123',
+    firstName: 'John',
+    lastName: 'Doe'
+  })
+);
+
+module.exports = Deserializer;
