--- test/__mocks__/jsonapi-serializer.js
@@ -1,26 +1,0 @@
-var Deserializer = jest.fn();
-var Serializer = jest.fn();
-
-Deserializer.prototype.deserialize = jest.fn().mockReturnValue(
-  Promise.resolve({
-    id: '123',
-    firstName: 'John',
-    lastName: 'Doe'
-  })
-);
-
-Serializer.prototype.serialize = jest.fn().mockReturnValue({
-  data: {
-    type: 'users',
-    id: '123',
-    attributes: {
-      'first-name': 'John',
-      'last-name': 'Doe'
-    }
-  }
-});
-
-module.exports = {
-  Serializer: Serializer,
-  Deserializer: Deserializer
-};
