+++ test/__mocks__/serializer.js
@@ -1,0 +1,14 @@
+var Serializer = jest.fn();
+
+Serializer.prototype.serialize = jest.fn().mockReturnValue({
+  data: {
+    type: 'users',
+    id: '123',
+    attributes: {
+      'first-name': 'John',
+      'last-name': 'Doe'
+    }
+  }
+});
+
+module.exports = Serializer;
